refactor: extract saga orchestration logic from gRPC services (task 12)
Move complex domain orchestration logic from grpc_service.go files into
dedicated orchestrator components:

- payment-order: Extract ~800 lines into payment_orchestrator.go
  - Saga pattern for payment execution with compensation
  - Lien management and gateway integration
  - Double-entry bookkeeping for ledger postings

- current-account: Extract ~350 lines into deposit_orchestrator.go
  - Deposit saga with position keeping and financial accounting
  - Double-entry bookkeeping with compensation on failure
  - Clearing account handling for debit/credit postings

The grpc_service.go files now focus on request validation and response
mapping, while orchestrators handle the business logic. All existing
tests updated to inject orchestrator dependencies.
