+++ services/payment-order/service/grpc_service_test.go
@@ -536,6 +536,18 @@ func testGatewayAccountConfig() *config.GatewayAccountConfig {
 	return cfg
 }
 
+// testOrchestrator creates a PaymentOrchestrator with the provided dependencies for testing.
+// This helper ensures tests that directly construct Service{} also get an orchestrator.
+func testOrchestrator(repo persistence.Repository, caClient CurrentAccountClient, faClient FinancialAccountingClient, gwConfig *config.GatewayAccountConfig) *PaymentOrchestrator {
+	return NewPaymentOrchestrator(PaymentOrchestratorConfig{
+		Logger:                    testLogger(),
+		Repo:                      repo,
+		CurrentAccountClient:      caClient,
+		FinancialAccountingClient: faClient,
+		GatewayAccountConfig:      gwConfig,
+	})
+}
+
 // Test InitiatePaymentOrder
 func TestInitiatePaymentOrder_Success(t *testing.T) {
 	repo := NewMockRepository()
@@ -833,13 +845,16 @@ func TestUpdatePaymentOrder_Settled(t *testing.T) {
 		},
 		executeLienDone: executeLienDone,
 	}
+	faClient := &MockFinancialAccountingClient{}
+	gwConfig := testGatewayAccountConfig()
 
 	svc := &Service{
 		repo:                      repo,
 		currentAccountClient:      caClient,
-		financialAccountingClient: &MockFinancialAccountingClient{},
-		gatewayAccountConfig:      testGatewayAccountConfig(),
+		financialAccountingClient: faClient,
+		gatewayAccountConfig:      gwConfig,
 		logger:                    testLogger(),
+		orchestrator:              testOrchestrator(repo, caClient, faClient, gwConfig),
 	}
 
 	// Create a payment order in EXECUTING state
@@ -882,13 +897,16 @@ func TestUpdatePaymentOrder_Settled_LienExecutionStatusTracking(t *testing.T) {
 		},
 		executeLienDone: executeLienDone,
 	}
+	faClient := &MockFinancialAccountingClient{}
+	gwConfig := testGatewayAccountConfig()
 
 	svc := &Service{
 		repo:                      repo,
 		currentAccountClient:      caClient,
-		financialAccountingClient: &MockFinancialAccountingClient{},
-		gatewayAccountConfig:      testGatewayAccountConfig(),
+		financialAccountingClient: faClient,
+		gatewayAccountConfig:      gwConfig,
 		logger:                    testLogger(),
+		orchestrator:              testOrchestrator(repo, caClient, faClient, gwConfig),
 	}
 
 	// Create a payment order in EXECUTING state
@@ -985,13 +1003,16 @@ func TestUpdatePaymentOrder_ByGatewayReferenceID(t *testing.T) {
 			},
 		},
 	}
+	faClient := &MockFinancialAccountingClient{}
+	gwConfig := testGatewayAccountConfig()
 
 	svc := &Service{
 		repo:                      repo,
 		currentAccountClient:      caClient,
-		financialAccountingClient: &MockFinancialAccountingClient{},
-		gatewayAccountConfig:      testGatewayAccountConfig(),
+		financialAccountingClient: faClient,
+		gatewayAccountConfig:      gwConfig,
 		logger:                    testLogger(),
+		orchestrator:              testOrchestrator(repo, caClient, faClient, gwConfig),
 	}
 
 	// Create a payment order in EXECUTING state
@@ -1054,12 +1075,15 @@ func TestUpdatePaymentOrder_Idempotent_Settled(t *testing.T) {
 	caClient := &MockCurrentAccountClient{
 		executeLienResp: &currentaccountv1.ExecuteLienResponse{},
 	}
+	faClient := &MockFinancialAccountingClient{}
+	gwConfig := testGatewayAccountConfig()
 	svc := &Service{
 		repo:                      repo,
 		currentAccountClient:      caClient,
-		financialAccountingClient: &MockFinancialAccountingClient{},
-		gatewayAccountConfig:      testGatewayAccountConfig(),
+		financialAccountingClient: faClient,
+		gatewayAccountConfig:      gwConfig,
 		logger:                    testLogger(),
+		orchestrator:              testOrchestrator(repo, caClient, faClient, gwConfig),
 	}
 
 	// Create a payment order in EXECUTING state
@@ -1699,6 +1723,14 @@ func TestSagaOrchestration_HappyPath(t *testing.T) {
 		},
 	}
 
+	// Create orchestrator with all dependencies
+	orchestrator := NewPaymentOrchestrator(PaymentOrchestratorConfig{
+		Logger:               slog.New(slog.NewJSONHandler(io.Discard, nil)),
+		Repo:                 repo,
+		CurrentAccountClient: caClient,
+		PaymentGateway:       gwMock,
+	})
+
 	// Create service with all dependencies configured
 	svc := &Service{
 		repo:                    repo,
@@ -1707,6 +1739,7 @@ func TestSagaOrchestration_HappyPath(t *testing.T) {
 		paymentGateway:          gwMock,
 		sagaTimeout:             DefaultSagaTimeout,
 		maxIdempotencyKeyLength: DefaultMaxIdempotencyKeyLength,
+		orchestrator:            orchestrator,
 		// kafkaProducer is nil - events won't be published but saga still runs
 	}
 
@@ -1759,6 +1792,14 @@ func TestSagaOrchestration_LienFailure(t *testing.T) {
 	// Gateway mock won't be called since lien fails first
 	gwMock := &MockPaymentGateway{}
 
+	// Create orchestrator with dependencies
+	orchestrator := NewPaymentOrchestrator(PaymentOrchestratorConfig{
+		Logger:               slog.New(slog.NewJSONHandler(io.Discard, nil)),
+		Repo:                 repo,
+		CurrentAccountClient: caClient,
+		PaymentGateway:       gwMock,
+	})
+
 	svc := &Service{
 		repo:                    repo,
 		logger:                  slog.New(slog.NewJSONHandler(io.Discard, nil)),
@@ -1766,6 +1807,7 @@ func TestSagaOrchestration_LienFailure(t *testing.T) {
 		paymentGateway:          gwMock,
 		sagaTimeout:             DefaultSagaTimeout,
 		maxIdempotencyKeyLength: DefaultMaxIdempotencyKeyLength,
+		orchestrator:            orchestrator,
 	}
 
 	ctx := context.Background()
@@ -1816,6 +1858,14 @@ func TestSagaOrchestration_GatewayFailure(t *testing.T) {
 		err: errGatewayUnavailable,
 	}
 
+	// Create orchestrator with dependencies
+	orchestrator := NewPaymentOrchestrator(PaymentOrchestratorConfig{
+		Logger:               slog.New(slog.NewJSONHandler(io.Discard, nil)),
+		Repo:                 repo,
+		CurrentAccountClient: caClient,
+		PaymentGateway:       gwMock,
+	})
+
 	svc := &Service{
 		repo:                    repo,
 		logger:                  slog.New(slog.NewJSONHandler(io.Discard, nil)),
@@ -1823,6 +1873,7 @@ func TestSagaOrchestration_GatewayFailure(t *testing.T) {
 		paymentGateway:          gwMock,
 		sagaTimeout:             DefaultSagaTimeout,
 		maxIdempotencyKeyLength: DefaultMaxIdempotencyKeyLength,
+		orchestrator:            orchestrator,
 	}
 
 	ctx := context.Background()
@@ -1899,6 +1950,14 @@ func TestSagaOrchestration_Timeout(t *testing.T) {
 
 	gwMock := &MockPaymentGateway{}
 
+	// Create orchestrator with dependencies
+	orchestrator := NewPaymentOrchestrator(PaymentOrchestratorConfig{
+		Logger:               slog.New(slog.NewJSONHandler(io.Discard, nil)),
+		Repo:                 repo,
+		CurrentAccountClient: caClient,
+		PaymentGateway:       gwMock,
+	})
+
 	// Configure very short saga timeout to trigger timeout
 	svc := &Service{
 		repo:                    repo,
@@ -1907,6 +1966,7 @@ func TestSagaOrchestration_Timeout(t *testing.T) {
 		paymentGateway:          gwMock,
 		sagaTimeout:             100 * time.Millisecond, // Short timeout
 		maxIdempotencyKeyLength: DefaultMaxIdempotencyKeyLength,
+		orchestrator:            orchestrator,
 	}
 
 	ctx := context.Background()
@@ -2050,6 +2110,14 @@ func TestSagaOrchestration_MalformedLienResponse(t *testing.T) {
 	}
 	gwClient := &MockPaymentGateway{}
 
+	// Create orchestrator with dependencies
+	orchestrator := NewPaymentOrchestrator(PaymentOrchestratorConfig{
+		Logger:               slog.New(slog.NewJSONHandler(io.Discard, nil)),
+		Repo:                 repo,
+		CurrentAccountClient: caClient,
+		PaymentGateway:       gwClient,
+	})
+
 	// Create service directly to avoid kafka producer requirement
 	svc := &Service{
 		repo:                    repo,
@@ -2058,6 +2126,7 @@ func TestSagaOrchestration_MalformedLienResponse(t *testing.T) {
 		paymentGateway:          gwClient,
 		sagaTimeout:             1 * time.Second,
 		maxIdempotencyKeyLength: DefaultMaxIdempotencyKeyLength,
+		orchestrator:            orchestrator,
 		// kafkaProducer is nil - events won't be published but saga still runs
 	}
 
@@ -2105,6 +2174,14 @@ func TestSagaOrchestration_GatewayPending(t *testing.T) {
 		},
 	}
 
+	// Create orchestrator with dependencies
+	orchestrator := NewPaymentOrchestrator(PaymentOrchestratorConfig{
+		Logger:               slog.New(slog.NewJSONHandler(io.Discard, nil)),
+		Repo:                 repo,
+		CurrentAccountClient: caClient,
+		PaymentGateway:       gwClient,
+	})
+
 	// Create service directly to avoid kafka producer requirement
 	svc := &Service{
 		repo:                    repo,
@@ -2113,6 +2190,7 @@ func TestSagaOrchestration_GatewayPending(t *testing.T) {
 		paymentGateway:          gwClient,
 		sagaTimeout:             5 * time.Second,
 		maxIdempotencyKeyLength: DefaultMaxIdempotencyKeyLength,
+		orchestrator:            orchestrator,
 		// kafkaProducer is nil - events won't be published but saga still runs
 	}
 
@@ -2185,6 +2263,8 @@ func TestUpdatePaymentOrder_LienExecutionFailure(t *testing.T) {
 		executeLienErr:  ErrLienServiceUnavailable,
 		executeLienDone: executeLienDone,
 	}
+	faClient := &MockFinancialAccountingClient{}
+	gwConfig := testGatewayAccountConfig()
 
 	// Use fast retry config for tests to avoid long wait times
 	fastRetryConfig := &clients.RetryConfig{
@@ -2197,10 +2277,11 @@ func TestUpdatePaymentOrder_LienExecutionFailure(t *testing.T) {
 	svc := &Service{
 		repo:                      repo,
 		currentAccountClient:      caClient,
-		financialAccountingClient: &MockFinancialAccountingClient{},
-		gatewayAccountConfig:      testGatewayAccountConfig(),
+		financialAccountingClient: faClient,
+		gatewayAccountConfig:      gwConfig,
 		logger:                    testLogger(),
 		lienExecutionRetryConfig:  fastRetryConfig,
+		orchestrator:              testOrchestrator(repo, caClient, faClient, gwConfig),
 	}
 
 	// Create a payment order in EXECUTING state
@@ -2328,13 +2409,16 @@ func TestPostLedgerEntries_FailureModes(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			repo := NewMockRepository()
+			caClient := &MockCurrentAccountClient{}
+			gwConfig := testGatewayAccountConfig()
 			svc := &Service{
 				repo:                      repo,
-				currentAccountClient:      &MockCurrentAccountClient{},
+				currentAccountClient:      caClient,
 				financialAccountingClient: tc.mockFA,
 				paymentGateway:            &MockPaymentGateway{},
-				gatewayAccountConfig:      testGatewayAccountConfig(),
+				gatewayAccountConfig:      gwConfig,
 				logger:                    testLogger(),
+				orchestrator:              testOrchestrator(repo, caClient, tc.mockFA, gwConfig),
 			}
 
 			// Create an executing payment order
